Update W for individual levels instead of a universal W for every level
The variable W is changed from an int to an array of ints "Ws" with length 32.
Ws[0] is the scaling parameter for level 0, Ws[1] is the scaling parameter for level 1, etc.
Read Amplification and Write Amplification are calculated using Ws[0].
If W would be replaced with candW, instead set Ws[0] to candW and update the other levels (Ws[0] becomes candW, Ws[1] becomes the previous Ws[0], Ws[2] becomes the previous Ws[1], ...)
If W would remain the same, keep current Ws[0] and update the other levels (Ws[0] remains unchanged, Ws[1] becomes the previous Ws[0], etc.)